requirements.txt
# Core ML/CV Libraries
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
pillow==10.0.0

# YOLO Detection
ultralytics==8.0.208
torch==2.0.1
torchvision==0.15.2

# OCR
easyocr==1.7.0
pytesseract==0.3.10

# Backend API
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3

# Database
psycopg2-binary==2.9.9

# Frontend
streamlit==1.28.1
streamlit-folium==0.13.0
folium==0.14.0
plotly==5.18.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.12.0
flake8==6.1.0
